third_party/WebKit/LayoutTests/fast/dom/MutationObserver/filesystem-callback-delivery-expected.txt

Test that Mutation Records are delivered following FileSystem callbacks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mutations.length is 1
PASS mutations[0].type is "attributes"
PASS mutations[0].attributeName is "foo"
PASS mutations.length is 1
PASS mutations[0].type is "attributes"
PASS mutations[0].attributeName is "baz"


PASS successfullyParsed is true

TEST COMPLETE
